Milliseconds method and Schedular partial implementation
