+++ README.md
@@ -14,3 +14,4 @@ MetaTime is licensed under the terms of the Apache v2.0 license.
 
 ## Authors
 * Tim Nieradzik
+* Syed Waqar Abbas
