+++ js/src/main/scala/pl/metastack/metatime/Platform.scala
@@ -1,0 +1,5 @@
+package pl.metastack.metatime
+
+object Platform {
+  implicit lazy val DefaultScheduler: Scheduler = new AsyncScheduler
+}
