--- shared/src/main/scala/pl/metastack/metatime/Period.scala
@@ -1,5 +1,0 @@
-package pl.metastack.metatime
-
-trait Period {
-
-}
