+++ shared/src/test/scala/pl/metastack/metatime/SchedulerTest.scala
@@ -1,33 +1,16 @@
 package pl.metastack.metatime
 
-import org.scalatest.FunSuite
+import org.scalatest.{AsyncFunSuite}
+import scala.concurrent.{Promise}
 
-class SchedulerTest extends FunSuite {
-  import Implicits._
+class SchedulerTest extends AsyncFunSuite  {
 
-  ignore("at()") {
-    val task = Scheduler.at(5.second.fromNow.component.asInstanceOf[DateTime]) {
-      ???
-    }
-
-    task.cancel()
-
-    val task2 = Scheduler.at(5.minutes.fromNow.component.asInstanceOf[DateTime]) {
-      ???
-    }
-  }
-
-  ignore("in()") {
-    val task = Scheduler.in(5.minutes) {
-      ???
-    }
-  }
-
-  ignore("every()") {
-    var counter = 0
-    val task = Scheduler.every(5.minutes) { case t =>
-      if (counter > 2) t.cancel()
-      counter += 1
-    }
+  test("at()") {
+    val scheduler: Scheduler = Platform.DefaultScheduler
+    val p = Promise[Unit]()
+      scheduler.at(Time(0, 0, 30, 0)) {
+        p.success(())
+      }
+    p.future.map (_ => assert(true))
   }
 }
